app.jsx cleaning, wand cursos separate file
