+++ MagicSchoolManagement/FrontendMagic/FrontendMagic/src/components/Cursor.css
@@ -1,0 +1,35 @@
+/* Cursor.css */
+* {
+    margin: 0;
+    cursor: none;
+}
+
+body {
+    background-color: #1a2a3a;
+    height: 100vh;
+    overflow: hidden;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+/* Wand handle (rectangle) */
+#cursor {
+    position: absolute;
+    width: 5px; /* Narrow rectangular handle */
+    height: 40px; /* Lengthen handle */
+    background-color: #a8e6cf;
+    transform: translate(-50%, -50%) rotate(-45deg); /* Rotate to point left */
+    pointer-events: none;
+}
+
+/* Starburst at the top of the wand */
+#cursor-star {
+    position: absolute;
+    width: 12px;
+    height: 12px;
+    background-color: #a8e6cf;
+    clip-path: polygon(50% 0%, 61% 35%, 98% 35%, 68% 57%, 79% 91%, 50% 70%, 21% 91%, 32% 57%, 2% 35%, 39% 35%);
+    transform: translate(-50%, -50%) rotate(-45deg); /* Rotate to match handle */
+    pointer-events: none;
+}
